Create Solidity: Create your token
This Solidity code defines a simple token contract named MyToken on the Ethereum blockchain, utilizing version 0.8.18 of the Solidity compiler. The contract specifies public variables for the token's name, "Cadbury" (TN), and its symbol, "CB" (TA). It also maintains a public counter, sale, which records the total number of tokens that have been sold or minted. A mapping called remaining tracks the balance of tokens for each address.

The mint function allows the creation of new tokens by increasing the sale counter by the specified amount and adding this amount to the balance of the specified address in the remaining mapping. Conversely, the burning function enables token consumption or destruction by reducing the sale counter and the address’s balance by the specified amount, but only if the address has enough tokens. This ensures that tokens cannot be burned if the address’s balance is insufficient. The functions in the contract allow for basic token minting and burning operations, managing the overall supply and individual balances of the token.